Failing test for [JAVA-57]
Failing test for https://datastax-oss.atlassian.net/browse/JAVA-57
Because I am not sure CCM runs on windows, did not use it in this test,
should be easy enough to convert it to a CCM test though.
This test assumes a keyspace named "test" exists. There is a create
table statement that can be commented in and out as well.