Duplicate Message classes from C* to work around classpath issues
Previously we relied on classpath ordering to ensure that the version
of org.apache.cassandra.transport.Message from the driver was loaded
instead of the one from C*. This is necessary as we backported some
message types from protocol v2. In client apps where its not easy to
control the classpath though, this is fragile so this patch modifies
the driver to move its version of Message to the
com.datastax.cassandra.transport namespace. This also requires us to
do the same with a number of classes which depend on Message.