Additional data migration time-based uuid methods
When converting from another data store, and when using time-based UUIDs as
part of the primary key, it was necessary to retroactively create valid
time-based uuids.

These new methods allow for that.  If this is useful to anyone, I thought
I'd push it early before I had time to go back and add proper unit tests.