Removed use of .stress.version file
It seems this file was only used so that we could precisely reference
the jar-with-dependencies. It can be replaced by a wildcard since we
only generate one such jar.

The .stress.version file can contain more than the version number in
case the JAVA_HOME environment variable is not correctly defined.
Therefore I think we can simply delete it instead of adding more
exclusion rules in the 'grep -Ev' command.