Temporary workaround to prevent ccm from switching to vnodes.
CASSANDRA-7649 exposed an issue with CCM: when a node which was running
without virtual tokens is restarted, it switches to virtual tokens. The
node now refuses to restart, which makes our tests fail.

The CCM issue will be fixed by https://github.com/pcmanus/ccm/pull/140.
In the meantime, this commit fixes the issue manually.