Revert "Ensure failed reconnection attempts don't re-trigger onDown."
This reverts commit f42f825d5014e3efedd54a77d715f9e696c26856.

This test is a bit fragile and could be harmful if we use unpooled
connections for something other than reconnection attempts in the
future. The only consequence is that onDown might be called multiple
times, which has a negligible impact (notifying the LBP multiple times,
which it's supposed to handle, and rescheduling a reconnection attempt,
which will not start upon encountering another one running).