Add tests for JAVA-543
This changed required to remove the "static" modifier from the
refreshSchema method.

Now that refreshSchema is an instance method, the Manager instance can
be retrieved from the local "cluster" field.  The isInitialConnection
parameter was unused and has been removed as well.

A utility method has been created in Keyspace so that table removal
operation are processed by that object for test purposes.