Update down after init test to bring down the first host tried
as a control connection.   This is needed to ensure that one host is
marked as down otherwise if the first host tried is up the down host
would not be marked as down until a Session is created.

The order of the hosts tried is dependent on the ordering of
metadata.hosts.values() which is JDK dependent.