Fix Integration Tests that were assuming 2 default core connections.
In 2.0.9 the default number of core connections increased from 2 to 8 as
a stopgap until JAVA-419 is fixed.  ClusterInitTest and SessionTest have
been updated to use the # of core connections from PoolingOptions to
factor in how many expected connections are attempted or opened.

sessionMemoryLeakTest creates 10,000 sessions and quite a bit of time with
8 connections per session.  Explicitly setting core connections to 2 will
reduce the amount of time and load it takes to run the test.