Update listener in comesUpWithin to also check for onAdd events.
This fixes PoolingOptionsTest#should_refresh_single_connected hosts.
This test was previously failing on jenkins as the Host is not always
marked up as soon as waitForUp completes as establishing the pool takes
just long enough in a constrained environment to cause the Host not to
be in an 'UP' state when comesUpWithin is called.  This would be fine
except that sometimes Cassandra inconsistently doesn't sent an 'UP'
topology status change event for new nodes (< 2.1 versions), so 'onUp'
is never called in the listener.  Checking for onAdd fixes this as well,
and we can be assured that the Host is actually up when this happens as
onAdd is invoked as soon as the pools are established.