Added OSGi example with Integration Test for testing different configs.
Includes a simple example of creating an OSGi bundle with a service that
uses driver-core.  Includes an Integration test that tests with the
following configs:

  * Default (default classifier with all dependencies)
  * Netty-Shaded (shaded classifier with all dependencies w/o netty)
  * Guava 15 (default classifier with Guava 15.0)
  * Guava 16
  * Guava 17
  * Guava 18

This test will if using the 'short' or 'long' profiles, so will be
pulled into continuous integration.

Uses pax-exam (https://ops4j1.jira.com/wiki/display/PAXEXAM4/Pax+Exam)
to test with different OSGi configurations.