Add driver-integration for testing with different package configs.
Since the behavior of the driver may be altered based on libraries that
are present we need a mechanism for testing with these configurations.
The 'osgi' example provides most of this functionality as it loads up
OSGi containers with different library configurations, but this not
adequate in some cases, particularly the 'shaded' case where some
classes use a different packaging namespace.