Add support for overriding PATH and JAVA_HOME used by ccm.
At times it is necessary to use a separate java install for CCM then
what is being used for running tests.  For example, if you want to run
tests with JDK 6 but against Cassandra 2.0, which requires JDK 7.

Added 2 new system properties 'ccm.path' which if provided will prepend
the value to PATH and 'ccm.java.home' which will set JAVA_HOME to the
value provided.