Fix memory leak caused by Connection#flusherLookup.
Discovered while running ClusterStressTest that there was a lot of
heap pressure as more Cluster instances were created.  Upon further
investigation found that Connection#flusherLookup was the culprit as it
is not cleaned up as Clusters are closed.  To solve this, made
flusherLookup a concurrent map with weak keys and made Flusher#eventLoop
a weak reference.  This way as the EventLoopGroups become closed their
associated EventLoops will be marked as weak references and eventually
cleaned up by the garbage collector.