Fix connection leak in borrowConnection when initing/closing.
Discovered that when a pool is replaced, borrowConnection may be called
on it before the initFuture has completed and the connections list is
populated.  This caused more connections to be created and associated
with the HostConnectionPool then there should have been.

Also check for 'isClosing' in borrowConnection instead of isClosed since
discardAvailableConnections() is called before the future is set, which
may cause borrowConnection() to create more connections while we are
closing.  This may not be an actual issue, but it seemed safe to do
this.