Add callback to close connection to handle cancellation of initFuture.
Additionally refactor Connection such that the initialization logic is
now moved into a separate method, initAsync.  This is parallel to the
changes in HostConnectionPool.

Reintroduce isInitialized to connection which is set to true if
initialization is successful.  Previously a connection may mistakenly be
marked as suspected if initFuture was cancelled and then an error occurs
on the connection.

Check for Pool being closed when initFuture completes as found a case
where Session and its Pools are closed, but identified a Pool where it
had Connections that were never closed.