Fix some timing and protocol version issues in Integration Tests.
Fixes timing issue in RecommissionedNodeTest by giving the Driver time
to mark a node down as it is additionally in an 'ADDED' state on connect
and will be marked down as soon as we detect invalid cluster
name/version.

Sets protocol version explicitly in
ReconnectionTest#should_use_connection_from_reconnection_in_pool so we
do not need to account for a possible reconnection(s) to demote the
protocol version.