Add DataType.allPrimitiveTypes(ProtocolVersion).
With the addition of tinyint, smallint, date and time we now have a case
where some primitive types are not supported in all native protocol
versions.  Therefore it seems useful to be able to request primitive
types by version.

This method is currently only exposed as package private for internal
use.  Updated tests to use this method in favor of
DataType.allPrimitiveTypes().