Record error when RetryDecision.tryNextHost.
JAVA-709 (#372) added the capability for a RetryDecision to specify that
a request may be retried on different host by using
RetryDecision.tryNextHost.  In the case where tryNextHost is used and
there are no more available hosts a NoHostAvailableException (no hosts
were tried) is surfaced to the user.   This change surfaces the
response error that initiated the RetryDecision in the raised
NoHostAvailableException's errors.

Updated tests that previously checked for UnavailableException to check
for NoHostAvailableException with an UnavailableException under error
for that host.