Refactor SchemaTest to make tests compatible with C* 2.2 + and JDK 8.
Some tests in the old SchemaTest class were failing due to the use of JDK 8 instead of JDK 6. For example, the order in which entries in a HashMap are printed is not the same, and some assertions were relying on a specific order.
Moreover, these tests were fragile they were relying on the equality between the CREATE TABLE statement and the string produced by TableMetadata.exportAsString().