Support cassandra.directory and cassandra.branch props for CCMBridge.
Adds optional cassandra.directory and cassandra.branch properties as
a means of running cassandra with CCM beyond deployed cassandra
versions.

The order of precedence for what is used is:

1. cassandra.directory
2. cassandra.branch
3. cassandra.version

cassandra.version is still required for @CassandraVersion annotation
checks for tests, but won't be used to launch CCM if either
cassandra.directory or cassandra.branch is present.