JAVA-922: Make TypeCodec return mutable collections.
This commit relaxes the postconditions on immutability stated in the javadocs for some methods in these classes.
This commit also updates the javadocs of GettableByIndexData and GettableByNameData accordingly.
It also prunes Row class and removes all duplicated method declarations.
The Upgrade guide is updated with behavioral changes introduced by this commit.