Fix Cluster.connect with a case-sensitive keyspace (JAVA-950).
Go back to the previous strategy of first initializing a session with no
keyspace and then issuing a USE statement. This guarantees that
SessionManager.poolsState.keyspace always contains a keyspace in the
"internal" format (unquoted and in the correct case), and keeps the
implementation simple.
The downside is that if the keyspace is wrong, we'll open all
connections before realizing the error, but that's an edge case.