JAVA-1009: use CL ONE to compute query plan when reconnecting control connection.
Before this change, it used the default CL which is a local one
(currently LOCAL_QUORUM, soon to be changed to LOCAL_ONE).

If the driver is configured to use DCAwareRoundRobinPolicy with
usedHostsPerRemoteDc > 0 but !allowRemoteDCsForLocalConsistencyLevel,
this would cause remote nodes to be excluded from the query plan.  If
the user specified usedHostsPerRemoteDc > 0 they probably expect that
remote hosts should be picked for the control connection regardless of
their default CL, hence the change.