Fetch the whole peers tables if not found specific host
We deploy cassandra in two datacenters(multi region). We use private ip as listen_address(also rpc_address) and public ip as broadcast_address. The public ip will change if instance restarts so we have to modify the broadcast_address. Cassandra cluster will handle it correctly but the driver will ignore the host on reconnection as no row is found in peer's table by using old public ip. In this case we can query the whole peers table as a fallback.