Revert "JAVA-995: Defunct connection even if initializing."
This reverts commit 192bd21688dbae236fa88fff61427bc7e35a3260.

This change caused another race condition, that could lead to marking
a host down if protocol negotiation failed at startup. We're reverting
to the previous behavior which is less severe (as explained in the
ticket, the only impact is that a host could be marked down later than
expected). The connection shutdown sequence will be reeaxamined to
provide a better fix.