Define isolated profile that disables tests at parent level.
The 'isolated' profile was introduced in driver-core a while back to run
tests in their own JVM fork when needed.  This causes problems however
when another submodule like driver-mapping doesn't have an isolated
profile in which it will run unit tests and overwrite any existing
surefire report.  This change disables tests when running with isolated
group unless you override it explicitly in a module that desires to use
it.