Suppress "Usage of API documented as @since 1.7+" IntelliJ errors.
IntelliJ can report inadequate usage of API call, which is very useful in normal situations,
but enabling this warning would mark the whole jdk8 package with errors.

This commit disables the errors for classes in this package, but to achieve this,
no imports could be used, and offending classes must appear with fully-qualified class names everywhere.