JAVA-1193: Refresh token and replica metadata synchronously when schema is altered
This commit also introduces the following refactorings:

1) Cache and reuse host and token information whenever possible

This commit modifies the TokenMap class so that it caches all
the information about hosts and tokens from the last token map
computation. This information can then be reused in a subsequent
token map computation, if only the schema has changed, but not
the topology.

2) Optimize calls to submitNodeListRefresh()

In some circonstances, this method could be called recursively
(i.e., triggered while already processing a node list refresh request).

This commit suppresses the calls to submitNodeListRefresh() from
ControlConnection and replaces them with calls to that method only
when appropriate, i.e.:
- When inside a node refresh, and the node has been added or removed;
- When reconnecting, and the node has been added.

The effect is that submitNodeListRefresh() is NOT called when in
triggerOnAdd() nor in triggerOnRemove(), because we are already inside
a node list refresh operation.

3) Minor optimizations

This commit optimizes code around TokenMap:
- Remove unnecessary field TokenMap.hosts
- Rename fields in TokenMap for consistency
- Avoid creating intermediary Set<String> instances by creating
Set<Token> instances directly when reading rows from system.local and
system.peers.