JAVA-1126: JAVA-1126: Handle schema changes in Mapper
The mapping manager subscribes to schema change events and reacts to
table and UDT removed / modified notifications: a warning is logged and,
for modifications, the manager tries to replace the mapper or codec with
a new one (which may or may not work depending on the nature of the
change and whether client code kept a reference to the old mapper
instance).

In addition, table/UDT names and column/field names are validated
earlier, in a fail-fast attempt to detect wrong configurations.