Allow to set default value when getting primitive type value is null
Motivation:

When we got primitive type such as calling getInt(int i),
we would get ‘0’ when the value is null. We can’t set the default value
if the value is null when expected the value isn't 0/false .
If we want to distinguish 0 and null , we should use isNull() or get(i,
Integer.class). It isn’t convenient for coding.
So add new methods to allow to setting the default value when got value
is Null.

Modifications:

add extra new methods with default value such as: 
public int getInt(int i, int defaultValue);

Result:

More flexible and convenient.