JAVA-1200: Upgrade LZ4 to 1.3.0.
This commit also upgrades Snappy version to 1.1.2.6 and Hdr Histogram version to 2.1.9.

Besides, it also introduces a new OSGi test to validate that LZ4 bundle can be used in an OSGi environment.
As this new test relies on a system property that might be present or not, this commit switches the Pax Exam container from "native" to "forked", because native containers are not confined enough to isolate system properties correctly.