JAVA-1200: Upgrade LZ4 to 1.3.0.
This commit also introduces a new OSGi test to validate that it can be
used in an OSGi environment.

As this new test relies on system properties that might be present or
not, this commit switches the Pax Exam container from "native" to
"forked", because native containers are not confined enough to isolate
system properties correctly.