JAVA-1196: Include hash of result set metadata in prepared statement id
Motivation:

See CASSANDRA-10786. When a table is altered (column is added), the wildcard queries
keep the old result metadata and never return the added column for all but the very
first client (as the very first client is the only one that receives UNPREPARED statement).
After adding result metadata to md5 on Cassandra side, as the message metadata on the client
is never updated and new client ID of the re-prepared message is not saved anywhere,
client would fall into the infinite re-prepare loop.

Modifications:

With Protocol V5:
- Include result metadata id in EXECUTE messages.
- Read result metadata id in PREPARED and ROWS responses.
- Replace cached prepared statements when result metadata id has changed.

Result:

Wildcard queries are now safe to use (see JAVA-420).
Result metadata id is now correctly handled.