JAVA-1212: Don't retry non-idempotent statements by default
This was opt-in behavior in 3.0.0, make it the default because it's the
right thing to do. The counterpart is that the isIdempotent flag needs
to be positioned properly on statements, the doc has been updated in
that regard.

This was made as an "internal" change (as opposed to a retry policy
implementation), so that it would work with any retry policy. Therefore
`IdempotenceAwareRetryPolicy` becomes redundant and was deprecated.