Eliminate memory copies in SnappyFrameCompressor and use netty's allocator for allocations
Motiviation:

Snappy allows to use ByteBuffer for compress/decompress when direct buffers are used. At the moment a memory copy to heap buffers is done, which is wasteful.

Modification:

Use ByteBuffer for compress/decompress when possible and allocate the buffer from the ByteBufAllocator of netty which may be a pooled implementation.

Result:

Less memory copies == Better performance.