JAVA-1261: Throw error when attempting to page in I/O thread
Iterating a result set past the current page triggers a blocking call
for the next page. This should never be done in a Netty I/O thread (for
example by iterating in a future callback), because it can deadlock.

Similar to JAVA-1182, we throw an error so that client code can detect
this mistake early.