JAVA-1283: Don't cache failed query preparations in the mapper
Motivation:

The object mapper prepares its queries asynchronously and caches the
resulting ListenableFuture instances. If a PREPARE call fails because of
a transient issue (e.g. all nodes are down), a failed ListenableFuture
will be cached.  Even after the issue gets resolved, the mapper will
rely on the failed future and keep throwing the same exception.

Modifications:

If preparation fails, do not keep the future in the cache.

Result:

The mapper will retry preparing the next time it's called, giving it a
chance to succeed if the issue was resolved.