JAVA-1355: Update getReplicas to handle range covering multiple hosts
Motivation:

Metadata.getReplicas(String, TokenRange) currently assumes that the
TokenRange provided represents a TokenRange for a single host.  Because
of this, it finds the replicas for the last Token on that range instead
of considering the entire range.

Arbitrary TokenRanges may be provided that span multiple hosts.  In
this case, this method should include each host spanning this range
and their replicas.

Modifications:

Updated method to evaluate which Host's TokenRanges intersect with the
input TokenRange and include those range's replicas.
Created method Metadata.getReplicas(String, Token) to query for
individual tokens.

Result:

Metadata.getReplicas(String, TokenRange) now includes replicas for all
hosts covering a range, not just the replicas for the last token in the
range.