JAVA-1362: Send query options flags as [int] for Protocol V5+
Motivation:

CASSANDRA-12838 changes the flags field on the QUERY, EXECUTE and BATCH
messages from [byte] to [int].  This change accounts for that change in
width in Protocol V5.

Modifications:

Write flags as int for Protocol V5+, byte otherwise.

Result:

QUERY, EXECUTE, and BATCH messages are now compatible with Protocol V5.