JAVA-1373: Encode flags in Prepare request for Protocol V5+
Motivation:

CASSANDRA-10145 introduces the capability to indicate the keyspace name
as part of QUERY, BATCH and PREPARE messages.  This commit simply
encodes empty flags for the PREPARE query to be compliant with Protocol
V5.

Modifications:

Write 4-byte int as 0 for empty flags in Prepare message for Protocol
Version 5+.

Result:

PREPARE messages are now compatible with Protocol V5.