Change SSLOptions API to take a remote address
Enable creation of SSLHandler with remote address information. This is
needed to instantiate SSLEngine/SSLHandler with remote address
information set, which is needed to enable hostname verification. We
cannot take this information from SocketChannel at this time since it is
not connected yet at the time Connection.Initializer creates the
SSLHandler.