JAVA-1367: Make protocol negotiation more resilient
Motivation:

CASSANDRA-11464 (introduced in 3.0.8+, 3.8+) changed logic of protocol
error messages to be sent with the protocol version requested by the
client, instead of a version that the server supports, to ensure that
the client will be able to decode the response properly.

The driver currently uses the returned protocol version to determine
what protocol version to downgrade to and reattempt connection.  Because
of this, the driver would retry the same protocol version and fail
again.

Modifications:

This commit changes the renegotiation algorithm to allow more than one
renegotiation attempt.  It also ignores the server response version and
always retries to connect using attempted version - 1.

This commit contains great contributions by Andrew Tolbert (@tolbertam).

Result:

Driver is compatible with Cassandra 3.0.8+, 3.8+.