JAVA-1364: Enable creation of SslHandler with remote address information
Motivation:

SSLOptions do not currently allow to create SslHandler instances with remote address information.
This is needed to instantiate SSLEngine/SslHandler with remote address information set,
which enables doing hostname verification (for example through a custom TrustManager).
We cannot take this information from SocketChannel since it is not connected yet
at the time Connection.Initializer creates the SslHandler.

Modifications:

Create a child interface to SSLOptions, RemoteEndpointAwareSSLOptions, that exposes a new
method to create SslHandler instances with remote address information.
Modify Connection class so that it detects instances of RemoteEndpointAwareSSLOptions
and calls the new method instead of the old one.
Create subclasses for JdkSSLOptions and NettySSLOptions that implement
RemoteEndpointAwareSSLOptions: RemoteEndpointAwareJdkSSLOptions and
RemoteEndpointAwareNettySSLOptions.
Deprecate SSLOptions and direct implementations.

Result:

Remote endpoint information is now available to implementors of RemoteEndpointAwareSSLOptions,
and to the driver built-in implementations RemoteEndpointAwareJdkSSLOptions and
RemoteEndpointAwareNettySSLOptions.

This commit is based on original work by Tim Lamballais Tessensohn (@wimtie).
It also contains contributions by Andrew Tolbert (@tolbertam) and
Olivier Michallat (@olim7t).