JAVA-1313: Copy SerialConsistencyLevel to PreparedStatement
Motivation:

A serial consistency level set on a statement that is going
to be prepared is not being propagated to the prepared statement.

Modifications:

Modify AbstractSession.prepareAsync(RegularStatement) and propagate
the serial consistency level to prepared statements, just as for
other inherited properties.

Result:

Serial consistency level is now being propagated to prepared
statements.