Add column instead of altering existing one in test
Motivation:

CASSANDRA-12443 removes support for 'ALTER TYPE' since it was deemed
unsafe.  Given that, tests that used it need to be changed.

Modifications:

Alter
SchemaChangesCCTest.should_receive_changes_made_while_control_connection_is_down_on_reconnect()
to instead of altering an existing column type to add a new column.
Since this test is validating that the schema is refreshed on reconnect
it is not significant that a column type was altered, but merely a
schema change is detected.

Added TableMetadataAssert.doesNotHaveColumn.

Result:

Test now passes against C* 3.10.