Support Duration type in PrimitiveTypeSamples. (#828)
* Support Duration type in PrimitiveTypeSamples.

A number of tests use PrimitiveTypeSamples for getting sample values for
given DataTypes.  Duration was previously omitted from this because it
is not supported in Set or Map keys.  Special case tests that involve
using these samples for Map and Sets so we get better Duration coverage.

In addition, this fixes a bug in UserTypesTest where it attempts to
resolve a sample for Duration but one is not found.