Fix intermittent ClusterInitTest failure
Nodes were being excluded from LoadBalancingPolicy because their
Cassandra version was too old, because of this nodes were never
marked down because we never attempted to connect to them.
Setting release_version to the minimum cassandra version for the
latest protocol version should prevent this from being an issue from
here on out.