JAVA-1499: Wait for load balancing policy at cluster initialization
Motivation:

In the previous version, the cluster was marked as initialized as soon
as the topology monitor had initialized. This created a race condition
where a session could be created before the LBP was initialized, observe
all nodes in the initial IGNORED state and not create any pool.

Modifications:

Don't complete the cluster init future until the LBP has initialized.

Result:

Cluster.connect is now guaranteed to see the node distances initialized
by the LBP.