Use id of UNPREPARED response to reprepare on the fly
Motivation:

When we execute a prepared statement and find out that the coordinator
doesn't have it in its cache, it's not always practical to rely on the
original statement to find out what to reprepare: in particular, it
could be batch and we would have to traverse all children to identify
the corresponding bound statement.

Modifications:

Use the id returned in the UNPREPARED response to look up the payload to
reprepare in a cache. Because we don't want to introduce a dependency
between CqlRequestHandler and CqlPrepareHandler, that cache is owned by
the Session, and contains only protocol-level data, in a class that is
agnostic to the PreparedStatement implementation.

Result:

Reprepare on the fly without inspecting the original statement. In
addition, having reprepare data at the session level will prove useful
if we decide to implement the "reprepare on up" feature.