JAVA-1599: exportAsString improvements (sort, format, clustering order)
JAVA-1407 improved exportAsString to ensure that user types were
described in topological order.  However a side effect of this change is
that User Types are not always consistently ordered as it depended on
undefined Map ordering behavior which may vary based on JDK
implementation.

While this is not a functional issue, it would be preferable that the
same order were always returned.  This implementation ensures this by
ordering the adjacency list of a vertex by alphabetical name of the user
type.  If there are no dependencies between user types, the result of
this change is that user types are ordered alphabetically.