JAVA-1682: Provide a way to record latencies for cancelled speculative executions
Those latencies are emitted with a
CancelledSpeculativeExecutionException, which allows LatencyTracker
implementations to choose whether to include them or not. For example, a
tracker that wants to detect "slowness" of a node might want to record
them (otherwise a node that always hits the speculative execution
threshold would appear to have no measurement), whereas a tracker that
monitors latencies of successful executions should ignore them.