JAVA-1858: Implement Serializable in implementations, not interfaces
Motivation:

A few of our types are serializable (row, data types, etc.), this is
mainly for integration with Spark. By making our interfaces extend
Serializable, we signal that all possible implementations should do it,
this is not necessarily the case if someone writes their own version in
an application that doesn't use Spark.

Modifications:

Remove `extends Serializable` from the interfaces, have the default
classes implement it instead.
Document in the interfaces that serializability is recommended, but not
strictly mandatory.

Result:

Less burden on implementors if they're not interested in Spark
integration.