JAVA-1872: Retain table's views when processing table update (#1035)
Motivation:

When a table update schema change event is received, the driver
creates a new TableMetadata instance for it and replaces the
existing one.

If that table has associated views, the driver currently does not
retain those views on the newly created TableMetadata instance.  This
change fixes that.

Modifications:

When a table update is received, update the base table on the views
tied to the old TableMetadata instance to the new one.  Added
MaterializedViewMetadata.setBaseTable to handle this.

Add SchemaChangesTest.should_retain_view_on_table_update to verify
this fix.

Result:

Views are now retained on table update schema events.