JAVA-1857: Add Statement.setHost
Motivation:

With the introduction of system views in Apache C* 4.0
(CASSANDRA-7622) there is a need for the capability to
set a target node when executing queries.

This feature should be avoided in the general case and
only used in exceptional cases such as this.

Modifications:

Add Statement.setHost(Host).

Update RequestHandler to bypass the configured load balancing policy
and instead create a single host query plan targeting the configured
host if set.

Result:

Driver now offers capability of host targeting via Statement.setHost.