JAVA-1903: Add WhiteListPolicy.fromHosts
Motivation:

Similarly to Cluster.Builder.addContactPoints(String ...) it would be
convenient for WhiteListPolicy to support a similar mechanism of
providing hostnames to resolve IP addresses from to be considered
for host selection.

Modifications:

Add WhiteListPolicy.fromHosts to create a WhiteListPolicy that uses
DNS to resolve IP addresses to be considered.

Result:

WhiteListPolicy can now be created from hostname(s).