JAVA-1771: Send driver name and version in STARTUP message
Motivation:

CASSANDRA-14275 adds the capability to parse driver identification
information from the STARTUP message in the form of DRIVER_NAME
and DRIVER_VERSION options.  This is useful for helping operators
understand what clients are connecting to Cassandra and what
drivers they are using.

Modifications:

DRIVER_NAME and DRIVER_VERSION configuration are now provided in
the STARTUP options map.  This is provided regardless of the
version of Cassandra we are connecting to, as this is not
a protocol version specific feature.

Result:

The driver will now send identifiying information in STARTUP
message on connection initialization.