JAVA-1942: Adapt type with CodecRegistry's associated AttachmentPoint
Motivation:

To ensure tuples and udts are decoded using the codec registry and
protocol version associated with the current Session, adapt any
incoming TupleValue / UdtValue's type to the AttachmentPoint of
the session when creating their associated codecs.

Modifications:

Update CachingCodecRegistry constructor to accept AttachmentPoint
to be associated with this registry.  Update DefaultDriverContext
to pass itself as the AttachmentPoint.

Update CachingCodecRegistry.createCodec methods to adapt types
to the AttachmentPoint.

Update DataTypeIT to create an AttachmentPoint based on the expected
ProtocolVersion to be used by the Session.

Introduce SessionUtils.getMaxProtocolVersion, which replaces
CassandraResourceRule.getHighestProtocolVersion.

Result:

Session's protocol version and codec registry are ensured to be
factored in when deserializing UDTs and Tuples.