JAVA-1211: Improve cluster init and shutdown (#1074)
Motivation:

To to improve messaging when a user attempts to use a Cluster that had previously been closed or failed to initialize.  In particular, this change addresses JAVA-1211, JAVA-1220 and JAVA-1929.

Modifications:

Adds additional null checks in Cluster.close so if a Cluster failed to completely initialize the resources that were created are closed (JAVA-1211).

Keeps reference to an Exception raised during Cluster initialization.  This is used to notify the user that the Cluster cannot be used because initialization failed and also prevents successive Cluster.init calls from trying to initialize the Cluster (JAVA-1220).

If Session is closed, preempt Session.execute queries with a specific exception indicating that the Session is closed (JAVA-1929).

Result:

User is now given more clear Exception messages if attempting to use or initialize a Cluster that had previously failed to initialize or was closed.  User is also provided more explicit Exception messages when attempting to execute queries against closed Sessions.