Include ServerError in downgrade criteria from peers_v2 to peers (#1085)
The logic from JAVA-1388 dictates that the driver downgrades from
system.peers_v2 to system.peers for node discovery if an
INVALID_QUERY error is received.  

DSE 6.0.0 to 6.0.2 with search enabled inappropriately responds with
a SERVER_ERROR with the message "Unknown keyspace/cf pair
(system.peers_v2)".  This commit handles this case specifically.