Generate empty javadoc JAR for the shaded core module
Motivation:

It looks like Maven Central has become less permissive, we used to
deploy the shaded core without a javadocs JAR but that's not accepted
anymore.

Modifications:

Generate a javadocs JAR during the build.

For now it's a dummy, empty JAR, users can refer to the javadocs of the
unshaded one. One downside is that this won't work in IDEs; if there's a
strong demand, we can look into generating the actual javadocs later.

Result:

The driver can be deployed into Maven central again.