JAVA-1995: Use rule chains to enforce JUnit rule order
Motivation:

Currently we don't control the execution order of the rules in our
integration tests. But in order to open a session, we need Cassandra to
be running, so we call cassandraResource.setUp() in
Sessionrule.before().

The problem is when this is coupled with a CassandraRequirement that
isn't met: cassandraResource.setUp() does not re-check the requirements,
so SessionRule forces the initialization of a cluster for a test that's
going to be ignored.

Modifications:

Use RuleChain to ensure that the server rule (CCM or Simulacron) always
initializes before the SessionRule.

Result:

No need to force the initialization of the server rule anymore.
Additionally, error handling in SessionRule.after() is simplified, since
the server rule is guaranteed to still be running.