Delay query plan computation until throttling ready
Motivation:

If a request is throttled, there will be a delay between the handler's
construction and the time it starts sending requests. So the handler
might operate on a query plan that is slightly out of date. Even though
the delay should be short, and handlers are resilient to stale plans, it
would be better to get the query plan as late as possible.

Modifications:

Build the query plan in onThrottleReady instead of the constructor. It
can't be a final field anymore, so pass it around as a method parameter
instead.

Result:

The query plan is built right before the first call to sendRequest.