Remove RequestHandler interface
Motivation:

The current API forces the concept of a "handler" on every
RequestProcessor implementation. This is not really necessary because
the driver does not manipulate this object other than directly calling
handle() on it.

Modifications:

Replace RequestProcessor.newHandler() by process(), which returns the
result directly -- in other words, the equivalent of
newHandler().handle().

Remove the RequestHandler interface.

Result:

Existing processors still use a separate handler (and this will probably
remain a common pattern) but they are not forced to do so.

In particular, it is now easier to implement a processor that caches its
results.