JAVA-2057: Ignore SUGGEST_UP TopologyEvent if pool is pending
Motivation:

There is a small window where a NEW_NODE event may be sent over the
control connection while a pool is initializing.  This is more likely to
happen in testing, where a cluster is brought up and the driver connects
to it immediately.

In this case, the driver would erroneously create another pool, which
would create additional connections.  The first pool created between
the currently initializing one and the new one being created would be
untracked with its connections remanining open even after the Session is
closed.

Modifications:

Akin to the logic in processDistanceEvent and processStateEvent, only
invoke createOrReconnectPool in onTopologyEvent when there is not an
initializing pool future present in the pending map.

Result:

createOrReconnectPool is no longer invoked in onTopologyEvent if there
is a pending initialization for that node.