JAVA-2058: Make programmatic config reloading part of the public API
Motivation:

Until now this was only possible via the internal API (by firing an
event on the bus). However this not such an exotic use case, it sounds
reasonable to expect that it will be commonly needed.

Modifications:

- add DriverConfigLoader.reload() and supportsReloading().
- remove ForceReloadConfigEvent, now obsolete since that was the only
  reason for its existence.

Result:

DriverConfigLoader is part of the public API and accessible through
DriverContext, so this does not require any internal imports anymore.