JAVA-2056: Reduce HashedWheelTimer tick duration
Motivation:
Timeouts and speculative executions are scheduled on the
same HashedWheelTimer, which has a default tick duration
of 100ms. This can be too imprecise for speculative executions.
Lowering the tick duration to 1ms does not affect performance,
but does allow for more precision for scheduling speculative
executions.

Modifications:
Reduce the default timer tick duration to 1ms (from 100ms)
and increase the default wheel size to 2048 (from 512).