JAVA-1612: Include netty-common jar in binary tarball
Motivation:

Users of the binary tarball noticed that the netty-common dependency was
missing, via a NoClassDefFoundError for io.netty.util.Timer.

This was caused by netty-common being excluded as it is a transitive
dependency of netty-transport-epoll.  While it is also a transitive
dependency of included jars, the maven-assembly-plugin still excludes
it.

Modifications:

Remove the use of useTransitiveFiltering in the dependencySet.  This is
ok as netty-transport-epoll dependencies are depended on by the driver jar
already.

Result:

netty-common jar is now included in the lib directory of the binary
tarball.