JAVA-2165: Abstract node connection information
Motivation:

Until now, the only information used to connect to a node was its socket
address (host+port). Internally, the driver also used this address as a
unique node identifier (it was the key in the Metadata.nodes map).
However, there are more complex deployment scenarios where the address
might not be sufficient: for example, if nodes are accessed through a
proxy that uses SNI routing, the connection information becomes
host+port+sni_server_name, and the host+port part is no longer unique.
We need to support those alternative connection methods, and use a
different unique node identifier.

Modifications:

- Introduce EndPoint as an abstract wrapper around connection
  information.  EndPoint.resolve() returns the socket address that will
  be used to open connections, but an endpoint may also contain
  additional pieces of information that will be used in other places
  (for example a custom SslHandlerFactory).
- Add Node.getEndPoint(), and modify DefaultTopologyMonitor to fill it.
  Allow contact points to be specified as EndPoints (preserving legacy
  methods).
- Replace InetSocketAddress by EndPoint everywhere that it was used to
  identify a node.
- Index Metadata.nodes by the true unique node identifier: host_id from
  system tables. This has two consequences:
  - we can't put contact points in Metadata.nodes directly, because we
    don't know their host_id yet. Store them separately
    (MetadataManager.contactPoints) and transfer them during the first
    node list refresh.
  - when we receive a status event, we can't use a simple map lookup
    anymore, so do a linear traversal to find the node with a matching
    broadcast_rpc_address (side effect: we don't need to translate it).

Result:

We no longer depend on nodes having a unique socket address.
We can plug in custom EndPoint implementations.