JAVA-2177: Don't exclude down nodes when initializing LBPs
Motivation:

In driver 4, the LBP "pushes" the distances to the driver, not the other
way around. So if we don't set the distance of a node in init(), it will
stay at IGNORED and the driver won't try to connect to it.
Therefore all nodes should be passed to init().

Modifications:

Don't filter the nodes in LoadBalancingPolicyWrapper.
Update the javadocs of LoadBalancingPolicy to explain the contract.
Update DefaultLoadBalancingPolicy to only add UP nodes to its live set.

Result:

Down nodes don't stay ignored anymore.