JAVA-2081: Fix reference configuration page
Abandon the idea of inlining the file's contents, this complicates our
documentation pipeline too much (it requires reStructuredText and
additional Python libraries).

Instead, link to the file in our GitHub repo. The link will have to be
updated before every release, but this already happens in other places
so we do a global search and replace.