JAVA-2255: Add annotation to indicate a property is transient
Motivation:

It is sometimes desirable for properties of an entity to not be
considered by an object mapper.

Modifications:

Introduced a @Transient annotation that when applied to a field or
getter, causes the mapper to not include this field for generated
queries.

In addition, if a field includes the 'transient' keyword, it is also not
considered.

Result:

properties including a @Transient annotation or having a field with
transient keyword will not be considered by the mapper.