JAVA-2254: Add support for computed properties
Motivation:

It is useful to map values that are the result of some Cassandra-side
computation, typically function calls, to properties.

Modifications:

Introduce a `@Computed` annotation that can be applied to a field or
getter that allows specifying a formula to instruct Cassandra to
compute and return a value back that can be mapped back to this
property.

Introduce AliasedPropertyDefinition for enabling use of aliases in query
specification.  This is needed because the resulting value name for a
function call is not deterministic.

Add EntityDefinition.getComputedValues() for retrieving computed values
and added getAllValues(), which is similar to getAllColumns() but also
returns computed values.

Add PropertyDefinition.getCqlResultName() to differentiate between
request name and result name for alias definitions.

Update EntityHelperGetMethodGenerator to allow selectively retrieving
values when their cql name does not match their result name.

Update EntityHelperSelectStartMethodGenerator to include computed values
in generated select query.

Result:

`@Computed` annotation may be used to specify properties that are based
on result of some function call.  This is only usable by
`@Select`-annotated dao methods.